backend/pyproject.toml
[project]
name = "saas-metering-platform"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "passlib[argon2]", 
  "python-jose[cryptography]",
  "python-multipart",
  "email-validator",
]

[tool.uvicorn]
factory = false
